# Vectorize Jpeg Encoder Color Conversion