@saucecontrol 2021-01-18
Rather than having two methods with a round-trip through memory, I'd try to get this all done in one step.  The best option would actually be to do the conversion fixed-point if AVX2 is available, but I assume that's a much larger change if the encoder is already taking the YCbCr input in floating point.

This version could be simplified using something like this:

```
// read 32 bytes, of which 24 will be used
rgb = vmovdqu(rgbptr)
RGBRGBRGBRGBRGBR|GBRGBRGBxxxxxxxx

// move 12 good bytes into each lane
rgb = vpermd(rgb, mask 0,1,2,6,3,4,5,7)
RGBRGBRGBRGBxxxx|RGBRGBRGBRGBxxxx

// group channel values together
rgb = vpshufb(rgb, mask 0,3,6,9,1,4,7,10,2,5,8,11,-1,-1,-1,-1)
RRRRGGGGBBBBxxxx|RRRRGGGGBBBBxxxx

// unpack to int32 width
rg = vpunpcklbw(rgb, zero)
R0R0R0R0G0G0G0G0|R0R0R0R0G0G0G0G0
bx = vpunpckhbw(rgb, zero)
B0B0B0B0x0x0x0x0|B0B0B0B0x0x0x0x0

r = vpunpcklwd(rg, zero)
R000R000R000R000|R000R000R000R000
g = vpunpckhwd(rg, zero)
G000G000G000G000|G000G000G000G000
b = vpunpcklwd(bx, zero)
B000B000B000B000|B000B000B000B000

// convert to float, then use the existing ConvertInternal math
// values will already be properly ordered for 32-byte stores
```

That's untested but should get you going in the right direction.  You'll need to allow for 8 bytes of overrun on the last read, or you can stop short and process the last 24 bytes using the LUT fallback or with the SIMD loop using a read/shift from the end.