@antonfirsov 2021-01-18
I'm super curious how much the algorithm in https://github.com/SixLabors/ImageSharp/pull/1508#issuecomment-761924703 could give us, but if there's no time to implement it, we can merge the PR as is after changing the ` Vector128.Create`-s.

In any case let's not close #1476 with the PR, there is definitely space left for improvements by rearranging the pipeline steps within the `Encode420` / `Encode444` loops. (Needs further analysis to determine what exactly has to be done.)