@tkp1n 2021-01-18
@antonfirsov I'm already on it.. Seems to reduce times roughly from 170ns to 100ns 🚀 (initial numbers)
@saucecontrol Thanks for the input!